9.2.6
Added a new page to store the generatePage func.  In this new file named page-template.js I cut and pasted in the func and added a line that exported the func like so.

+const generatePage = (name, github) => {
    return `
    <!DOCTYPE html>
    <html lang="en">
    <head>
      <meta charset="UTF-8">
      <meta name="viewport" content="width=device-width, initial-scale=1.0">
      <meta http-equiv="X-UA-Compatible" content="ie=edge">
      <title>Portfolio Demo</title>
    </head>

    <body>
      <h1>${name}</h1>
      <h2><a href="https://github.com/${github}">Github</a></h2>
    </body>
    </html>
    `;
};

module.exports = generatePage;

then in the app.js where the func was removed from I added a require and set it to a variable that has the same name as the function so it doesn't mess with the code I wrote previously in the fs.writeFile function.

const generatePage = require('./src/page-template.js');