9.3.4
used npm init and hit enter several times to get through the prompts.  It created a file called {}package.json.  Now that we have access to npm we went ahead and installed one of it's modules called inquirer using the terminal npm install inquirer.  There is good info describing what and why we are doing this though we have not been intructed to add this to app.js just yet.