9.3.5
This was so much info.  We added questions using inquirer.prompt.  We made a call that used .then to call multiple functions in order.  Then we also allowed the function to to run the questions again if selected and pushed it all into an array.  I will essentially need to reread this page.  Useful stuff but painfully dense.

.then(projectData => {
      portfolioData.projects.push(projectData);
pushes all the info into the array
      if (projectData.confirmAddProject) {
        return promptProject(portfolioData);
then listens to the name: confirmAddProject and if true runs this function again with the data just created passed in
      } else {
        return portfolioData;
otherwise leaves the function with the data just created
      }
    });