9.3.6
Added validates to make sure the user does not accidentally skip required fields.  Like so
 validate: githubInput => {
        if (githubInput) {
          return true;
        } else {
          console.log('Please enter your GitHub username!');
          return false;
        }
      }

then added this
 {
      type: 'confirm',
      name: 'confirmAbout',
      message: 'Would you like to enter some information about yourself for an "About" section?',
      default: true
to ask if you want to add info about yourself.  Then we attach a when statement that references this to only display a field to type this in only if it is true.
 message: 'Provide some information about yourself:',
      when: ({ confirmAbout }) => {
        if (confirmAbout) {
          return true;
        } else {
          return false;
        }
      }