9.4.4
updated HTML and inserted them data into the file looking something like this.  ${header.name} etc......

<h1 class="page-title text-secondary bg-dark py-2 px-3">${
          header.name
        }</h1>
        <nav class="flex-row">
          <a class="ml-2 my-1 px-2 py-1 bg-secondary text-dark" href="https://github.com/${
            header.github
          }">GitHub</a>

then we went into the app.jss and updated the call to actually write the file with

 fs.writeFile("./index.html", pageHTML, (err) => {
      if (err) throw new Error(err);
The first part names it index.html, the second part catches the object that gets sent back and names it pageHTML???? and the third handles it's error.

None of this works if
const fs = require("fs");
is not enabled up top.