9.4.5
created a function that checks if the about section exists and if it does creates a HTML section to add to the document.

const generateAbout = (aboutText) => {
  if (!aboutText) {
    return "";
if there is no about section it returns an empty string.
  }

  return `
    <section class="my-3" id="about">
      <h2 class="text-dark bg-primary p-2 display-inline-block">About Me</h2>
      <p>${aboutText}</p>
    </section>
  `;
};

Then in the actual HTML we call it like this

<main class="container my-5">
    ${generateAbout(about)}
    </main>
by passing in the data detructured objects about section.