9.4.6
used .map and .filter to check to see if we selected feature and added that.  Pretty dense stuff and I will have to read this module again.  There was also a pretty huge refractor that took the code from something easy to read to something a little crazy.  Again I will have to come back to this one a read it.