9.5.6
added the calls to the bottom of generate-site.js so it looks like this
module.exports = { writeFile, copyFile };

and then deleted

const fs = require("fs");
because we are no longer using fs in app.js and added

const { writeFile, copyFile } = require("./utils/generate-site.js");
to used object destructuring to take the to items and use them in app.js.

This also provides a very useful step by step overview of how the entire site works.